fix: return total elements in pagination

+++ src/main/java/com/tekton/challenge/model/dto/PaginationDto.java
@@ -1,5 +1,5 @@
 package com.tekton.challenge.model.dto;
 
-public record PaginationDto(Integer totalElements, Integer totalPages) {
+public record PaginationDto(Long totalElements, Integer totalPages) {
 
 }
